Added setting parameters
Added setting parameters: settings to be recoreded once, and save as
parameters.

They can be specified with the add_parameters command, which now takes
up to two inputs: parameters, and settings. Their defaults are [].
Parameters is as before.

Settings is a list of the form [ entry, entry, ... ] where entries are
either of the form [VDS_ID,vds_name,parameter_name] or
[[server,device,setting],inputs,parameter_name].

parameter_name is optional and if left out a name will be given to the
parameter of the form "server (device) setting(inputs)"