Added UI files, updated readme, added blank run_ui.py
Added UI files from qtdesigner, and a batch file for compiling them.
This batch file will be updated as more UI files are added, to compile
all UI files in the folder.
Updated readme.md to describe the divisions of functionality within the
repository (sweeper, UI, connecting code)
Added run_ui.py, currently blank. This is the file that will connect the
sweeper to the UI, and run the UI.