Added sweep check for setting written as parameters
Added sweep check that ensures that not all of the recoreded settings
are written as parameters.
Recorded settings that are set to ignore all present axes will be
measured once at sweep start and written as parameters; therefore, they
do not count as having an active recorded setting. To make sure that
there is at least one active recorded setting, not all of the recorded
settings may ignore all of the axes.